feat: move articles page to the separate component
